Add files via upload
Robot.java is working as intended and fulfills all the assignment requirements, as far as I can tell. That said, there is some vaguery in the assignment parameters- for example, it says to "direct robot to pickup at x,y," when the robot is not *at* x,y, but method pickUp() is not meant to move the robot. Also I initialized the grid with '.' instead of spaces because it makes the output actually readable. We can clear these minor things up with the instructor later, but I'd hate to sacrifice the opportunity for the code to be good just to make it right. Any improvements or additions are welcome, I just wanted to start us off with what I had so far.

Robot.java is independent of App.java, RobotController.java and RoboGUI.fxml. I have done very little with the GUI stuff so far, but the files are all linked, and running it launches Scene Builder. I suspect Robot.java will need to be merged into the FXML stuff at some point but I don't yet know enough about the interface stuff to say.